+++ crates/action-graph/src/action_graph_builder.rs
@@ -319,7 +319,6 @@ impl<'app> ActionGraphBuilder<'app> {
             persistent: task.is_persistent(),
             runtime: self.get_runtime(project, task.platform, true),
             target: task.target.to_owned(),
-            timeout: task.options.timeout,
             id: None,
         });
