+++ crates/process/src/lib.rs
@@ -5,6 +5,7 @@ mod output;
 mod process_error;
 mod shell;
 
+pub use async_command::*;
 pub use command::*;
 pub use moon_args as args;
 pub use output::*;
